feat(js): export core types to TypeScript, add LoxoneMiniserver (#83)
Make the library usable from plain TypeScript/JavaScript projects without a
JS-only facade layer. Core Kotlin types are annotated with @JsExport directly
so TypeScript consumers access Control, LoxoneState, ValueState etc. as-is.

Breaking changes (warranted — BC incompatibilities were explicitly accepted):
- LoxoneState is no longer suspend-based; @Volatile copy-on-write replaces Mutex
- LoxoneState.getAllUuids() returns Array<String> instead of Set<String>
- WeatherEvent/WeatherState.lastUpdate: UInt → Int (UInt is not JS-exportable)
- Renamed LoxoneSession → LoxoneMiniserver (entity over lifecycle)

New:
- LoxoneMiniserver in commonMain: high-level entry point for all targets (JVM, native, JS)
  - Primary constructor takes LoxoneEndpoint; URL-string secondary for convenience
- LoxoneEndpoint exported to TypeScript with @JsExport; companion factory methods
  (local, public, fromUrl) exposed as true statics via @JsStatic
- JsLoxoneMiniserver in jsMain: @JsExport subclass exported as "LoxoneMiniserver"
  - Primary constructor: new LoxoneMiniserver(endpoint, user, pass)
  - URL factory: LoxoneMiniserver.fromUrl("http://...", user, pass)
  - onValueChanged/onTextChanged callbacks and Promise-returning lifecycle wrappers
- Array convenience getters on @JsExport domain types for KtList/KtMap fields:
  LoxoneApp.controlsArray/roomsArray/catsArray/mediaServerArray,
  Control.subControlsArray/linksArray, Statistic.outputsArray, StatisticV2.groupsArray,
  StatisticGroup.dataPointsArray, WeatherServer.weatherFieldTypesArray,
  DaytimerState/WeatherState.entriesArray
- ES module library output with TypeScript definitions
  (binaries.library + generateTypeScriptDefinitions)
- Long exported as bigint via -Xes-long-as-bigint compiler flag
- docs/js-interop-design.md documenting the non-obvious design decisions

+++ src/commonMain/kotlin/Codec.kt
@@ -278,7 +278,7 @@ object Codec {
 
         while (buffer.remaining() >= WeatherEvent.HEADER_SIZE_BYTES) {
             val uuid = readUuid(buffer)
-            val lastUpdate = buffer.readUnsignedInt()
+            val lastUpdate = buffer.readUnsignedInt().toInt()
             val entryCount = buffer.readInt()
 
             if (entryCount < 0 || buffer.remaining().toLong() < entryCount.toLong() * WeatherEntry.SIZE_BYTES) break
